Build binary search tree
My JavaScript file ios now set up with a node factory function, a tree
factory function, and a build tree function, all of which work together
to take in an ordered array of integers to create a binary search tree
of node objects. The array is also ordered prior to all of this using
the mergeSort and removeDuplicates functions. Finally, the created
binary search tree is printed to the console and therefor tested using
the prettyPrint variable function, which was provided as a way to test
the tree by the Odin Project.