Assign new tree object to variable
Prior to this commit, my tree function was not a proper factory
function. Now, the tree factory function contains the buildTree as a
method, and returns the root of the tree for it to be used in the
prettyPrint function. I also started to write methods in the returned
tree object for insertion and deletion of nodes.