Edit insert method to use find method
Prior to this commit, In my insert method, I realized that my first
check of checking whether that value exists or not was using the passed
in array. It should instead search the tree manually, because once I add
to the tree, I am not adding to the array. This is therefor flawed. The
insert method now uses the find method to verify whether the data exists
in the tree prior to insertion.