Create isBalanced method for Tree factory function
This method checks to see if the binary search tree is balanced. I also
has to modify the factory function slightly by defining the returned
object as "tree", and then returning tree. This is because within my
isBalanced method, I needed to call the find method, which the call its
self was is found within a treeTraversal helper function. The scope of
this was off, and I needed a way to refer to the find method. Now, when
calling the find method in this circumstance, I use tree.find().